requirements.txt
mysql-connector-python >= 9.5.0
python-dotenv >= 1.1.1
flask >= 8.1.3